赛前代码

+++ CodeCraft-2019/lib/scheduler.h
@@ -4,6 +4,6 @@
 
 bool Scheduler(Map &cityMap);
 
-int getA(Map &cityMap);
+double getA(Map &cityMap);
 
 #endif
